+++ src/address.js
@@ -43,7 +43,7 @@ function toBech32(data, version, prefix) {
 exports.toBech32 = toBech32;
 function fromOutputScript(output, network) {
   // TODO: Network
-  network = network || networks.bitcoin;
+  network = network || networks.liquid;
   try {
     return payments.p2pkh({ output, network }).address;
   } catch (e) {}
@@ -60,7 +60,7 @@ function fromOutputScript(output, network) {
 }
 exports.fromOutputScript = fromOutputScript;
 function toOutputScript(address, network) {
-  network = network || networks.bitcoin;
+  network = network || networks.liquid;
   let decodeBase58;
   let decodeBech32;
   try {