+++ src/payments/p2ms.js
@@ -42,7 +42,7 @@ function p2ms(a, opts) {
     },
     a,
   );
-  const network = a.network || networks_1.bitcoin;
+  const network = a.network || networks_1.liquid;
   const o = { network };
   let chunks = [];
   let decoded = false;