+++ test/ts-node-register.js
@@ -1,5 +1,5 @@
 // This file is required to run mocha tests on the TS files directly
 
 require("ts-node").register({
-  project: "test/tsconfig.json",
+    project: "test/tsconfig.json",
 });