+++ ts_src/ecpair.ts
@@ -56,7 +56,7 @@ class ECPair implements ECPairInterface {
     if (options === undefined) options = {};
     this.compressed =
       options.compressed === undefined ? true : options.compressed;
-    this.network = options.network || NETWORKS.bitcoin;
+    this.network = options.network || NETWORKS.liquid;
 
     if (__Q !== undefined) this.__Q = ecc.pointCompress(__Q, this.compressed);
   }
@@ -132,7 +132,7 @@ function fromWIF(wifString: string, network?: Network | Network[]): ECPair {
 
     // otherwise, assume a network object (or default to bitcoin)
   } else {
-    network = network || NETWORKS.bitcoin;
+    network = network || NETWORKS.liquid;
 
     if (version !== (network as Network).wif)
       throw new Error('Invalid network version');