+++ ts_src/networks.ts
@@ -7,6 +7,7 @@ export interface Network {
   pubKeyHash: number;
   scriptHash: number;
   wif: number;
+  assetHash: string;
 }
 
 interface Bip32 {
@@ -14,36 +15,27 @@ interface Bip32 {
   private: number;
 }
 
-export const bitcoin: Network = {
-  messagePrefix: '\x18Bitcoin Signed Message:\n',
-  bech32: 'bc',
+export const liquid: Network = {
+  messagePrefix: '\x18Liquid Signed Message:\n',
+  bech32: 'lq',
   bip32: {
     public: 0x0488b21e,
     private: 0x0488ade4,
   },
-  pubKeyHash: 0x00,
-  scriptHash: 0x05,
+  pubKeyHash: 57,
+  scriptHash: 39,
   wif: 0x80,
+  assetHash: '6f0279e9ed041c3d710a9f57d0c02928416460c4b722ae3457a11eec381c526d',
 };
 export const regtest: Network = {
-  messagePrefix: '\x18Bitcoin Signed Message:\n',
-  bech32: 'bcrt',
-  bip32: {
-    public: 0x043587cf,
-    private: 0x04358394,
-  },
-  pubKeyHash: 0x6f,
-  scriptHash: 0xc4,
-  wif: 0xef,
-};
-export const testnet: Network = {
-  messagePrefix: '\x18Bitcoin Signed Message:\n',
+  messagePrefix: '\x18Liquid Signed Message:\n',
   bech32: 'tb',
   bip32: {
     public: 0x043587cf,
     private: 0x04358394,
   },
-  pubKeyHash: 0x6f,
-  scriptHash: 0xc4,
+  pubKeyHash: 235,
+  scriptHash: 75,
   wif: 0xef,
+  assetHash: '5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225',
 };