+++ ts_src/payments/p2sh.ts
@@ -1,5 +1,5 @@
 import * as bcrypto from '../crypto';
-import { bitcoin as BITCOIN_NETWORK } from '../networks';
+import { liquid as BITCOIN_NETWORK } from '../networks';
 import * as bscript from '../script';
 import {
   Payment,