+++ ts_src/transaction_builder.ts
@@ -138,7 +138,7 @@ export class TransactionBuilder {
   // WARNING: maximumFeeRate is __NOT__ to be relied on,
   //          it's just another potential safety mechanism (safety in-depth)
   constructor(
-    public network: Network = networks.bitcoin,
+    public network: Network = networks.liquid,
     public maximumFeeRate: number = 2500,
   ) {
     this.__PREV_TX_SET = {};